requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.8.1
django-on-heroku==1.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
postmarker==0.18.2
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
